development-requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
cffi==1.15.0
colorama==0.4.4
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
whitenoise==6.2.0
